Raised the warning level on MSVC8 to 4, and fixed several warnings caused by that.
Originally committed to SVN as r2045.

+++ aegisub/main.cpp
@@ -117,7 +117,7 @@ void SetThreadName(DWORD dwThreadID, LPCSTR szThreadName) {
 // Gets called when application starts, creates MainFrame
 bool AegisubApp::OnInit() {
 #ifdef __VISUALC__
-	SetThreadName(-1,"AegiMain");
+	SetThreadName((DWORD) -1,"AegiMain");
 #endif
 
 	StartupLog(_T("Inside OnInit"));
