+++ aegisub/vfw_wrap.cpp
@@ -78,7 +78,7 @@ wxArrayInt VFWWrapper::GetKeyFrames(wxString filename) {
 
 	// Loop through stream
 	for (size_t i=0;i<frame_c;i++) {
-		if (AVIStreamIsKeyFrame(ppavi,i)) {
+		if (AVIStreamIsKeyFrame(ppavi,(int)i)) {
 			frames.Add(i);
 		}
 	}
